feat: Add Two Factor authentication view and route
This commit adds a new view and route for Two Factor authentication in the app_ecommerce project. The view is implemented in the TwoFactorView.vue file and the route is defined in the router/index.ts file. This feature allows users to authenticate using a two-factor code in addition to their email and password.

Co-authored-by: Carlos
Co-authored-by: Fer